requirements.txt
requests>=2.18.4, <2.22.0
selenium>=3.14.0
pika>=0.12.0, <1.0.0
python-hosts>=0.4.5
apscheduler>=3.5.3
dnspython>=1.15.0
psutil>=5.4.8
netifaces>=0.10.7
distro>=1.3.0
pycurl>=7.43.0.2
python-crontab>=2.3.6
enum34>=1.1.6           # Python 3.4 Enum backported to 3.3, 3.2, 3.1, 2.7, 2.6, 2.5, and 2.4
paho-mqtt==1.4.0

